backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==1.3.0
anthropic==0.7.0
google-generativeai==0.3.2
requests==2.31.0
PyYAML==6.0.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
Pillow==10.0.1
numpy==1.24.3

# GCP 서비스
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.38.0
vertexai==0.0.1

# 벡터 검색 및 임베딩
faiss-cpu==1.7.4
chromadb==0.4.15
sentence-transformers==2.2.2

# 웹 크롤링 및 파싱
selenium==4.15.0
webdriver-manager==4.0.1

# 개발 도구 (선택사항)
pytest==7.4.3
flake8==6.1.0
black==23.11.0
mypy==1.7.1 